backend/pyproject.toml
[tool.poetry]
name = "lintchecker"
version = "0.1.0"
description = ""
authors = ["Ivan Popov <ma1n.py@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ruff = "0.0.287"
asgiref = "3.7.2"
attrs = "23.1.0"
build = "0.10.0"
cachecontrol = "0.13.1"
certifi = "2023.7.22"
cffi = "1.15.1"
charset-normalizer = "3.2.0"
cleo = "2.0.1"
crashtest = "0.4.1"
cryptography = "41.0.3"
distlib = "0.3.7"
django = "4.2.5"
django-rest-framework = "0.1.0"
djangorestframework = "3.14.0"
dulwich = "0.21.6"
filelock = "3.12.3"
idna = "3.4"
importlib-metadata = "6.8.0"
installer = "0.7.0"
jaraco-classes = "3.3.0"
jeepney = "0.8.0"
jsonschema = "4.17.3"
keyring = "24.2.0"
more-itertools = "10.1.0"
msgpack = "1.0.5"
packaging = "23.1"
pexpect = "4.8.0"
pkginfo = "1.9.6"
platformdirs = "3.10.0"
poetry = "1.6.1"
poetry-core = "1.7.0"
poetry-plugin-export = "1.5.0"
ptyprocess = "0.7.0"
pycparser = "2.21"
pyproject-hooks = "1.0.0"
pyrsistent = "0.19.3"
pytz = "2023.3.post1"
rapidfuzz = "2.15.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
secretstorage = "3.3.3"
shellingham = "1.5.3"
sqlparse = "0.4.4"
tomlkit = "0.12.1"
trove-classifiers = "2023.8.7"
urllib3 = "2.0.4"
virtualenv = "20.24.5"
zipp = "3.16.2"
flake8 = "^6.1.0"
celery = "^5.3.4"
flower = "^2.0.1"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"
django-debug-toolbar = "^4.2.0"
drf-yasg = "^1.21.7"
types-requests = "^2.31.0.2"
environs = "^9.5.0"
pika = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
